+++ deepcave/runs/run.py
@@ -17,14 +17,15 @@ from pathlib import Path
 import ConfigSpace
 import jsonlines
 import numpy as np
-from ConfigSpace.configuration_space import Configuration
-from ConfigSpace.read_and_write import json as cs_json
+from ConfigSpace.configuration_space import Configuration, ConfigurationSpace
+from ConfigSpace.hyperparameters.hp_components import ROUND_PLACES
 
 from deepcave.runs import AbstractRun, Status, Trial
 from deepcave.runs.objective import Objective
 from deepcave.utils.compression import Encoder
 from deepcave.utils.files import make_dirs
 from deepcave.utils.hash import string_to_hash
+from deepcave.utils.util import config_to_tuple
 
 
 class Run(AbstractRun, ABC):
@@ -287,6 +288,8 @@ class Run(AbstractRun, ABC):
             config_id_len = len(self.configs)
             self.configs[config_id_len] = config
             self.origins[config_id_len] = origin
+            # Use same rounding as ConfigSpace does
+            self.config_id_mapping[config_to_tuple(config, ROUND_PLACES)] = config_id_len
 
         config_id = self.get_config_id(config)
         if config_id is None:
@@ -350,7 +353,7 @@ class Run(AbstractRun, ABC):
         self.path = Path(path)
 
         # Save configspace
-        self.configspace_fn.write_text(cs_json.write(self.configspace))
+        self.configspace.to_json(self.configspace_fn)
 
         # Save meta data (could be changed)
         self.meta_fn.write_text(json.dumps(self.meta, cls=Encoder, indent=4))
@@ -417,7 +420,7 @@ class Run(AbstractRun, ABC):
         self.meta = json.loads(self.meta_fn.read_text())
 
         # Load configspace
-        self.configspace = cs_json.read(self.configspace_fn.read_text())
+        self.configspace = ConfigurationSpace.from_json(self.configspace_fn)
 
         # Load configs
         configs = json.loads(self.configs_fn.read_text())
@@ -427,6 +430,13 @@ class Run(AbstractRun, ABC):
         # Load origins
         origins = json.loads(self.origins_fn.read_text())
         self.origins = {int(k): v for k, v in origins.items()}
+
+        # Make sure there is a config_id_mapping
+        if not self.config_id_mapping:
+            for config_id, config in self.configs.items():
+                if isinstance(config, Configuration):
+                    config = config.get_dictionary()
+                self.config_id_mapping[config_to_tuple(config, ROUND_PLACES)] = config_id
 
         # Load history
         with jsonlines.open(self.history_fn) as f: