+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Converters
 - Add support for AMLTK.
 
+## Version-Updates
+- ConfigSpace from 0.6.1 to 1.2.0
+- Numpy from 1.26.4 to 2.0.1
+
+## Ablation Paths
+- Added ablation paths as a plugin.
+- Added ablation as a evaluator to use for the plugin.
+- Added tests as well as documentation.
+- Modified the RandomForest surrogate model, so it can be passed the number of trees.
+
 # Version 1.2.1
 
 ## Quality of Life